refactor: relocate tournament commands/handlers/validators to `Domain` layer and update imports
- Moved `CreateTournamentValidator`, `DeleteTournamentCommand`, `CreateTournamentCommand`, and related handlers to the `Domain` layer.
- Updated import paths throughout the affected files for consistency.
- Adjusted error handling in `CreateTournamentValidator` to use `InternalBackendApiErrorVerifyUserExists`.
